Изменение на long

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        int[] sales = {10,45,87,56,90,23,100,23,110,80};
+        long[] sales = {10,45,87,56,90,23,100,23,110,80};
         SalesManager salesManager = new SalesManager(sales);
         System.out.println(" МАКСИМАЛЬНАЯ ПРОДАЖА  "+ salesManager.max() + " РУБ.");
     }
